Update on the heating element controller.

+++ io.openems.edge.controller.HeatingElementController/test/io/openems/edge/controller/HeatingElementController/HeatingElementTest.java
@@ -84,7 +84,7 @@ public class HeatingElementTest {
 		}
 
 		@Override
-		public int minTime() {
+		public double minTime() {
 			return minTime;
 		}
